refactor: refactor this feature into a tool
