+++ src/extension/tools/index.ts
@@ -3,6 +3,7 @@ import { ElementClick } from './element_click';
 import { ExportFile } from './export_file';
 import { ExtractContent } from './extract_content';
 import { FindElementPosition } from './find_element_position';
+import { GetAllTabs } from './get_all_tabs';
 import { OpenUrl } from './open_url';
 import { Screenshot } from './screenshot';
 import { TabManagement } from './tab_management';
@@ -16,6 +17,7 @@ export {
   ExtractContent,
   FindElementPosition,
   OpenUrl,
+  GetAllTabs,
   Screenshot,
   TabManagement,
   WebSearch,
